# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: CareerStack AI API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /profile:
    get:
      operationId: getProfile
      tags: [profile]
      summary: Get current user profile
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "404":
          description: Profile not found
    post:
      operationId: createProfile
      tags: [profile]
      summary: Create user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProfileBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
    patch:
      operationId: updateProfile
      tags: [profile]
      summary: Update user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfileBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
  /projects:
    get:
      operationId: listProjects
      tags: [projects]
      summary: List all projects for current user
      responses:
        "200":
          description: List of projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"
    post:
      operationId: createProject
      tags: [projects]
      summary: Create a new project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProjectBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
  /projects/{id}:
    get:
      operationId: getProject
      tags: [projects]
      summary: Get a project by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Project details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "404":
          description: Not found
    patch:
      operationId: updateProject
      tags: [projects]
      summary: Update a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProjectBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
    delete:
      operationId: deleteProject
      tags: [projects]
      summary: Delete a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /roadmaps:
    get:
      operationId: listRoadmaps
      tags: [roadmaps]
      summary: List all roadmaps for user
      responses:
        "200":
          description: Roadmaps
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Roadmap"
    post:
      operationId: generateRoadmap
      tags: [roadmaps]
      summary: Generate a roadmap for a technology
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerateRoadmapBody"
      responses:
        "201":
          description: Roadmap generated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Roadmap"
  /roadmaps/{id}:
    get:
      operationId: getRoadmap
      tags: [roadmaps]
      summary: Get roadmap details with milestones
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Roadmap with milestones
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RoadmapDetail"
        "404":
          description: Not found
  /roadmaps/{roadmapId}/tasks/{taskId}/toggle:
    patch:
      operationId: toggleTask
      tags: [roadmaps]
      summary: Toggle task completion status
      parameters:
        - name: roadmapId
          in: path
          required: true
          schema:
            type: integer
        - name: taskId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Task updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
...
components:
  schemas:
...
    UserProfile:
      type: object
      properties:
        id:
          type: integer
        clerkId:
          type: string
        name:
          type: string
        email:
          type: string
        college:
          type: ["string", "null"]
        degree:
          type: ["string", "null"]
        graduationYear:
          type: ["integer", "null"]
        careerGoal:
          type: ["string", "null"]
        preferredDomain:
          type: ["string", "null"]
        interests:
          type: array
          items:
            type: string
        skills:
          type: array
          items:
            type: string
        profilePhotoUrl:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - clerkId
        - name
        - email
        - interests
        - skills
        - createdAt
        - updatedAt
    CreateProfileBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        college:
          type: string
        degree:
          type: string
        graduationYear:
          type: integer
        careerGoal:
          type: string
        preferredDomain:
          type: string
        interests:
          type: array
          items:
            type: string
        skills:
          type: array
          items:
            type: string
      required:
        - name
        - email
    UpdateProfileBody:
      type: object
      properties:
        name:
          type: string
        college:
          type: string
        degree:
          type: string
        graduationYear:
          type: integer
        careerGoal:
          type: string
        preferredDomain:
          type: string
        interests:
          type: array
          items:
            type: string
        skills:
          type: array
          items:
            type: string
        profilePhotoUrl:
          type: string
    Project:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        problemSolved:
          type: ["string", "null"]
        technologies:
          type: array
          items:
            type: string
        difficultyLevel:
          type: string
          enum: [beginner, intermediate, advanced]
        githubLink:
          type: ["string", "null"]
        liveLink:
          type: ["string", "null"]
        screenshotUrl:
          type: ["string", "null"]
        duration:
          type: ["string", "null"]
        role:
          type: ["string", "null"]
        category:
          type: ["string", "null"]
        completionStatus:
          type: string
          enum: [planning, in_progress, completed]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - title
        - technologies
        - difficultyLevel
        - completionStatus
        - createdAt
        - updatedAt
    CreateProjectBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        problemSolved:
          type: string
        technologies:
          type: array
          items:
            type: string
        difficultyLevel:
          type: string
          enum: [beginner, intermediate, advanced]
        githubLink:
          type: string
        liveLink:
          type: string
        screenshotUrl:
          type: string
        duration:
          type: string
        role:
          type: string
        category:
          type: string
        completionStatus:
          type: string
          enum: [planning, in_progress, completed]
      required:
        - title
        - technologies
        - difficultyLevel
        - completionStatus
    UpdateProjectBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        problemSolved:
          type: string
        technologies:
          type: array
          items:
            type: string
        difficultyLevel:
          type: string
          enum: [beginner, intermediate, advanced]
        githubLink:
          type: string
        liveLink:
          type: string
        screenshotUrl:
          type: string
        duration:
          type: string
        role:
          type: string
        category:
          type: string
        completionStatus:
          type: string
          enum: [planning, in_progress, completed]
...
    Roadmap:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        technology:
          type: string
        totalMilestones:
          type: integer
        completedMilestones:
          type: integer
        progressPercent:
          type: number
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - technology
        - totalMilestones
        - completedMilestones
        - progressPercent
        - createdAt
    RoadmapDetail:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        technology:
          type: string
        milestones:
          type: array
          items:
            $ref: "#/components/schemas/Milestone"
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - technology
        - milestones
        - createdAt
    Milestone:
      type: object
      properties:
        id:
          type: integer
        roadmapId:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        orderIndex:
          type: integer
        estimatedDuration:
          type: string
        industryRelevance:
          type: string
        status:
          type: string
          enum: [not_started, in_progress, completed]
        tasks:
          type: array
          items:
            $ref: "#/components/schemas/Task"
      required:
        - id
        - roadmapId
        - title
        - orderIndex
        - estimatedDuration
        - industryRelevance
        - status
        - tasks
    Task:
      type: object
      properties:
        id:
          type: integer
        milestoneId:
          type: integer
        title:
          type: string
        completed:
          type: boolean
      required:
        - id
        - milestoneId
        - title
        - completed
    GenerateRoadmapBody:
      type: object
      properties:
        technology:
          type: string
      required:
        - technology
...