# lib/api-zod/src/generated/types/roadmapDetail.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CareerStack AI API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Milestone } from "./milestone";

export interface RoadmapDetail {
  id: number;
  userId: number;
  technology: string;
  milestones: Milestone[];
  createdAt: Date;
}