# lib/api-zod/src/generated/types/updateProjectBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CareerStack AI API specification
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateProjectBodyCompletionStatus } from "./updateProjectBodyCompletionStatus";
import type { UpdateProjectBodyDifficultyLevel } from "./updateProjectBodyDifficultyLevel";

export interface UpdateProjectBody {
  title?: string;
  description?: string;
  problemSolved?: string;
  technologies?: string[];
  difficultyLevel?: UpdateProjectBodyDifficultyLevel;
  githubLink?: string;
  liveLink?: string;
  screenshotUrl?: string;
  duration?: string;
  role?: string;
  category?: string;
  completionStatus?: UpdateProjectBodyCompletionStatus;
}